requirements.txt
Django==2.2.24
djangorestframework==3.12.4
psycopg2-binary==2.8.6
pytz==2021.3
sqlparse==0.4.2
